requirements.txt
pydantic
ipykernel
python-dotenv
numpy
pandas
matplotlib
seaborn
flask
memory_profiler
streamlit
langchain
langsmith
langchain_community
chromadb
pypdf
beautifulsoup4
selenium
arxiv
pydantic-settings
pymupdf
langchain-text-splitters
langchain_openai
faiss-cpu
langchain_chroma
langchain_community
langchain_groq
fastapi
uvicorn
langserve
sse_starlette